Trait diem_scratchpad::ProofRead

pub trait ProofRead<V>: Sync {
    fn get_proof(&self, key: HashValue) -> Option<&SparseMerkleProof<V>>;
}

A type that implements ProofRead can provide proof for keys in persistent storage.

Required Methods

fn get_proof(&self, key: HashValue) -> Option<&SparseMerkleProof<V>>

Gets verified proof for this key in persistent storage.